+++ structs.go
@@ -2,7 +2,7 @@ package main
 
 // User struct info contains data fro 1 row of users table
 type User struct {
-	ID       int64
+	ID       int32
 	Name     string
 	Contact  string
 	Contact2 string
@@ -10,8 +10,8 @@ type User struct {
 
 // Item struct info contains data fro 1 row of items table
 type Item struct {
-	ID            int64
+	ID            int32
 	Name          string
 	Description   string
-	CurrentUserID int64
+	CurrentUserID int32
 }
